Use shared examples for OIDC and SAML LOA3 specs
**Why**: To make it easier to remember that LOA3 features should always
be tested with both SAML and OIDC

This is a small PR to show what I have in mind for refactoring these
tests and to make sure we have consistent coverage between SAML and
OIDC.
